Koushi_Proj/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
async-timeout==4.0.2
attrs==22.2.0
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.3
datasets==2.10.1
decorator==5.1.1
dill==0.3.6
entrypoints==0.4
filelock==3.9.0
frozenlist==1.3.3
fsspec==2023.3.0
gitdb==4.0.10
GitPython==3.1.31
googletrans==3.1.0a0
gTTS==2.3.1
h11==0.9.0
h2==3.2.0
happytransformer==2.4.1
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.13.1
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.0.0
Jinja2==3.1.2
jsonschema==4.17.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
multidict==6.0.4
multiprocess==0.70.14
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.0
pandas==1.5.3
pdf2image==1.16.3
Pillow==9.4.0
protobuf==3.20.3
pyarrow==11.0.0
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
PyPDF2==3.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
responses==0.18.0
rfc3986==1.5.0
rich==13.3.1
semver==2.13.0
sentencepiece==0.1.97
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
streamlit==1.19.0
streamlit-player==0.1.5
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
torch==1.13.1
tornado==6.2
tqdm==4.65.0
transformers==4.27.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
Unidecode==1.3.6
urllib3==1.26.14
validators==0.20.0
watchdog==2.3.1
xxhash==3.2.0
yarl==1.8.2
zipp==3.15.0
